Interactive threshold editing. Unfinished drop station function. New calc.numbers module for soem useful calculations.
Main current issue with the stations editor is to get changes in the class object t o probagate down to the source object. Had trouble doing this using the xarray.drop functionality, so I tried del - but hard to figure out exactly how to do this. May haev to do some kind of global (re) definition of D within the class?